requirements.txt
flask
pymongo[srv]
gunicorn
python-dotenv
click==7.1.2
dnspython==1.16.0
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pymongo==3.11.3
Werkzeug==1.0.1